Put check_call back in errors
